Crash on exceptions
Prior to this change all exception were caught. That was okay when the
cookie was static and the only way to recover from an error was retry.
Now that cookies can be refresh by relogging in the best course of
action in a failure is to crash and let the OS restart us, triggering a
new login.